extrai Register em um componente RegisterUserForm
